Massive re-write of strava script
Adds support to fully authenticate users and provide detailed
statistics for detected activites. If a user haven't authorized
access to the data they will be prompted to do so upon new activity
detection. Once they do so (if they use the provided link) the activity
will be announced again in full detail.