Update getLatestTrack to support current response format
Previously a single track object was returned when there was no
"now playing" info available. It now always seems to be an array
irregardless of now playing tracks.

However, the now playing tracks doesn't seem to be reported at all.
At least this is the case with Spotify reported tracks.